Add HTML boilerplate, webpage title and metadata to index.html
I added the HTML boilerplate so that I could add the essential information and sections for a webpage to index.html. I added a title so that users can recognize the main topic of the webpage from the web broswer tab and the metadata to provide information about the contents or topic of my webpage so that search engines can improve the ranking of my webpage.