Add an icon to index.html
I added an icon to the webpage in order to assist in differentiating my webpage's browser tab from other webpages that are open on tabs on the user's web browser. Additionally, I added the same "about-me-transparent-PNG" icon with different file extensions so that the web browser could display the icon using the first icon with a file extension that the web browser supports.