Create a form in my-restaurant.html and add username field
I created a form in my-restaurant.html so that users can enter and submit the necessary information to book a table at My Restaurant. I added the username field to the form and required customers to enter a username with a minimum of two characters so that customers can utilize a valid username to register at the restaurant.