Style the main-content, sidebar, and navigation sections of index.html
I added a grid layout to the main-content and sidebar sections of the Flexbox and Grid Practice webpage so that the porportion of these sections to the entire width of the viewport can remain as accurate as possible and the spacing between these sections can remain generally stable even if the size of the viewport is adjusted. I adjusted the padding of the navigation section so that the size of the header and footer containers can be similar.